+++ pylinac/acr.py
@@ -160,7 +160,7 @@ class SpatialResolutionModule(CatPhanModule):
 
     def _setup_rois(self) -> None:
         for name, setting in self.roi_settings.items():
-            self.rois[name] = HighContrastDiskROI(
+            self.rois[name] = HighContrastDiskROI.from_phantom_center(
                 self.image,
                 setting["angle_corrected"],
                 setting["radius_pixels"],
@@ -552,7 +552,7 @@ class MRSlice11PositionModule(CatPhanModule):
     def _setup_rois(self) -> None:
         for name, setting in self.roi_settings.items():
             # angle is +90 because pointing right is 0, and these rois move downward, not rightward
-            self.rois[name] = RectangleROI(
+            self.rois[name] = RectangleROI.from_phantom_center(
                 self.image,
                 setting["width_pixels"],
                 setting["height_pixels"],
@@ -639,7 +639,7 @@ class MRSlice1Module(CatPhanModule):
             )
         # spatial res
         for name, setting in self.roi_settings.items():
-            self.rois[name] = HighContrastDiskROI(
+            self.rois[name] = HighContrastDiskROI.from_phantom_center(
                 self.image,
                 setting["angle_corrected"],
                 setting["radius_pixels"],
@@ -754,7 +754,7 @@ class MRUniformityModule(CatPhanModule):
     def _setup_rois(self) -> None:
         super()._setup_rois()
         for name, roi in self.ghost_roi_settings.items():
-            self.ghost_rois[name] = RectangleROI(
+            self.ghost_rois[name] = RectangleROI.from_phantom_center(
                 self.image,
                 roi["width_pixels"],
                 roi["height_pixels"],